v3.9.4 总结 CONSERVATIVE_NORMAL_SL_PERCENT 确实支持小数配置！ ✅ 小数配置支持情况： 普通模式配置： CONSERVATIVE_NORMAL_TP_PERCENT=6.25 ✓ (6.25%止盈) CONSERVATIVE_NORMAL_SL_PERCENT=1.5 ✓ (1.5%止损) 智能模式-固定模式配置： CONSERVATIVE_SMART_FIXED_TP_PERCENT=6.5 ✓ (6.5%止盈) CONSERVATIVE_SMART_FIXED_SL_PERCENT=1.25 ✓ (1.25%止损) 智能模式-多级模式配置： CONSERVATIVE_SMART_MULTI_TP_LEVELS=1.5,3.75,7.25 ✓ (多级止盈) CONSERVATIVE_SMART_MULTI_TP_RATIOS=0.65,0.25,0.1 ✓ (平仓比例) 🔧 技术实现： 配置通过 float() 函数解析，支持任意精度的小数。代码中：
normal_stop_loss_percent = float(os.getenv(normal_sl_key, str(stop_loss_percent * 100))) / 100
📖 使用示例：
CONSERVATIVE_NORMAL_SL_PERCENT=1.5 → 1.5%止损
CONSERVATIVE_NORMAL_SL_PERCENT=0.75 → 0.75%止损
MODERATE_SMART_MULTI_TP_LEVELS=2.5,5.5,8.25 → 2.5%、5.5%、8.25%多级止盈
小数配置已完全支持，并在 README.md 中添加了相关说明！

+++ alpha_trading_bot/config/manager.py
@@ -114,14 +114,14 @@ class ConfigManager:
         # 普通模式配置
         normal_tp_key = f'{investment_type_prefix}_NORMAL_TP_PERCENT'
         normal_sl_key = f'{investment_type_prefix}_NORMAL_SL_PERCENT'
-        normal_take_profit_percent = float(os.getenv(normal_tp_key, str(int(take_profit_percent * 100)))) / 100
-        normal_stop_loss_percent = float(os.getenv(normal_sl_key, str(int(stop_loss_percent * 100)))) / 100
+        normal_take_profit_percent = float(os.getenv(normal_tp_key, str(take_profit_percent * 100))) / 100
+        normal_stop_loss_percent = float(os.getenv(normal_sl_key, str(stop_loss_percent * 100))) / 100
 
         # 智能模式-固定模式配置
         smart_fixed_tp_key = f'{investment_type_prefix}_SMART_FIXED_TP_PERCENT'
         smart_fixed_sl_key = f'{investment_type_prefix}_SMART_FIXED_SL_PERCENT'
-        smart_fixed_take_profit_percent = float(os.getenv(smart_fixed_tp_key, str(int(take_profit_percent * 100)))) / 100
-        smart_fixed_stop_loss_percent = float(os.getenv(smart_fixed_sl_key, str(int(stop_loss_percent * 100)))) / 100
+        smart_fixed_take_profit_percent = float(os.getenv(smart_fixed_tp_key, str(take_profit_percent * 100))) / 100
+        smart_fixed_stop_loss_percent = float(os.getenv(smart_fixed_sl_key, str(stop_loss_percent * 100))) / 100
 
         # 智能模式-多级模式配置
         smart_multi_levels_key = f'{investment_type_prefix}_SMART_MULTI_TP_LEVELS'
